backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
Pillow
